backend/requirements/base.txt
django==2.0.2
djangorestframework==3.7.7
django-filter==1.1.0
dj-database-url==0.5.0
pillow==5.0.0
python-dateutil==2.6.1
whitenoise==3.3.1
boto==2.48.0

django-rest-auth==0.9.2
facebookads==2.11.4
facebook_business==3.2.2
